feat: add collection components
